-DAQs are reset upon starting NR, to remove any tasks that might be holding resources NR could need. -'input settings' (sampling rate, A2D gain) moved to the HardwareSettings window. -merge with Jon's commit from earlier today.
git-svn-id: https://neurorighter.googlecode.com/svn/trunk@547 471701b0-e017-11dd-9759-139c9ab14fc6

+++ NeuroRighter/Filters/MUAFilter.cs
@@ -28,7 +28,7 @@ namespace NeuroRighter.Filters
         private int BufferLength;
         private int NumChannels;
         private List<NationalInstruments.Analysis.Dsp.Filters.ButterworthLowpassFilter> LowFilters;
-        private int samplingRate;
+        private double samplingRate;
 
         private const double WINDOW = 5.0; //in seconds, how much data to average over for RMS
         private const double VOLTAGE_EPSILON = 1E-6;
@@ -37,7 +37,7 @@ namespace NeuroRighter.Filters
         private readonly int numReadsPerWindow;
 
 
-        internal MUAFilter(int NumChannels, int InputSamplingRate, int BufferLength, double highCut,  int filterOrder,
+        internal MUAFilter(int NumChannels, double InputSamplingRate, int BufferLength, double highCut,  int filterOrder,
             int DownsampleFactor, double DeviceRefresh)
         {
             this.NumChannels = NumChannels;
